Separate lib_machine tests into separate file
